Fix ref parsing in 'git br' alias
refname:short gives a non-ambiguous name, but the ambiguity is among all refs in
the repository not just the refs specified to the command.

When there are tags with the same same as branches then the 'br' vs 'branch -vv'
output is different because the refname:short gives 'heads/branch' instead of
'branch'.

refname:lstrip=-1 does the expected thing, and makes the output match again.