Ubuntu 20.04 updates
PATH security hole still present.

If the user's bin directories don't exist and then an app is installed that
uses them, the PATH will not be set correctly. Change the philospohy to always
add the user's bin directories to the PATH and create the directories if they
don't exist. This allows newly installed binaries to be active immediately.