Add btprofile()
This facilitates switching Bluetooth profiles on headphones where A2DP is
usually in use for music, but HFP is needed when the microphone is required.

module-bluetooth-policy (pacmd list-modules) is loaded and works to switch
profiles when an audio client is registered as media=phone. Google Meet doesn't
appear to do that.

Not sure if Windows is doing something a bit smarter or what.