Add a reminder to add a swap target
I'd like to be able to toggle HFP and A2DP since that's my common used
case. Encapsulating the command that determines the current profile has
to deal with multiple-levels-of-shell-escaping hell which I don't have
time for right now.

And just duplicating that code is inelegant and prone to
forgot-to-update-every-location bugs when the tools that manipulate
bluetooth inevitably change.