Sort-of revert Vim 8's defaults handling
Reported the issue to tpope, turns out I'd misdiagnosed the exact
problem. Still the defaults file but it was 'filetype plugin indent on'
that was the issue, not ttimeout/len.

And tpope fixed the breakage so fugitive once again works in all cases.
Thanks, tpope!

Update my comments to account for the new situation.