Add comments on CRLF settings
It's been a while and I've forgotten the details, which, for EOL
handling means things may go wrong.

And in fact they did. :-(

A checkin in this repo was failed due to 'safecrlf = true' after being
checked out under Git For Window's bash environment. This has prompted
switching to using .gitattributes instead.