Remove "!git" anti-pattern
The first alias was added in a batch with the rest of the .gitconfig and
I can't recall now why I might have done that.

I can't see it helping at all, and it's a bit slower to re-invoke git a
second time, so remove the "!git" anti-patterns.

Tested that this doesn't break the "git root" alias.